fix: allow reconnecting with the same peer id
This handles the case when one of the peers was disconnected without others knowing it,
ex. iframe was reloaded from the inside.
If another party continues to listen -> it will accept the handshakes with the same id by default.

one.listen('two');
two.connect('one');
two.connect('one'); <-- this will re-establish connection